pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["jyoh5"]
license = "MIT"
readme = "README.md"
packages = [{include = "airbnb_clone_backend"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.7"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
pyjwt = "^2.6.0"
django-environ = "^0.10.0"
django-cors-headers = "^3.14.0"
requests = "^2.28.2"
dj-database-url = "^1.3.0"
psycopg2-binary = "^2.9.6"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
gunicorn = "^20.1.0"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
